Refactor (split) widgets into containers, labels and buttons
